+++ .github/workflows/build.yml
@@ -41,7 +41,7 @@ jobs:
 
           - os: windows-latest
             job-name: "Windows Tests"
-            gradle-tasks: "jvmTest"
+            gradle-tasks: "jvmTest -x :integration-test:jvmTest"
             max-workers: 3
 
           - os: macos-latest
@@ -64,6 +64,11 @@ jobs:
         with:
           node-version: 'lts/*'
 
+      - name: Set up Node Packages
+        if: matrix.os == 'ubuntu-latest'
+        working-directory: integration-test/src/jvmTest/typescript
+        run: npm ci
+
       - name: Setup Gradle
         uses: gradle/actions/setup-gradle@v5
         with:
@@ -75,7 +80,7 @@ jobs:
             caches
             notifications
             sdks
-            .konan/**
+            ../.konan/**
 
       - name: ${{ matrix.job-name }}
         run: |-
@@ -144,6 +149,11 @@ jobs:
         with:
           add-job-summary: 'always'
           cache-read-only: true
+          gradle-home-cache-includes: |
+            caches
+            notifications
+            sdks
+            ../.konan/**
 
       - name: Download Maven Local Repository
         uses: actions/download-artifact@v7
