+++ integration-test/src/jvmTest/kotlin/io/modelcontextprotocol/kotlin/sdk/integration/typescript/AbstractKotlinClientTsServerTest.kt
@@ -17,7 +17,7 @@ abstract class AbstractKotlinClientTsServerTest : TsTestBase() {
     protected abstract suspend fun <T> useClient(block: suspend (Client) -> T): T
 
     @Test
-    @Timeout(30, unit = TimeUnit.SECONDS)
+    @Timeout(20, unit = TimeUnit.SECONDS)
     fun connectsAndPings() = runBlocking(Dispatchers.IO) {
         useClient { client ->
             assertNotNull(client, "Client should be initialized")
@@ -30,7 +30,7 @@ abstract class AbstractKotlinClientTsServerTest : TsTestBase() {
     }
 
     @Test
-    @Timeout(30, unit = TimeUnit.SECONDS)
+    @Timeout(20, unit = TimeUnit.SECONDS)
     fun listsTools() = runBlocking(Dispatchers.IO) {
         useClient { client ->
             val result = client.listTools()
@@ -44,7 +44,7 @@ abstract class AbstractKotlinClientTsServerTest : TsTestBase() {
     }
 
     @Test
-    @Timeout(30, unit = TimeUnit.SECONDS)
+    @Timeout(20, unit = TimeUnit.SECONDS)
     fun callGreet() = runBlocking(Dispatchers.IO) {
         useClient { client ->
             val testName = "TestUser"
@@ -59,7 +59,7 @@ abstract class AbstractKotlinClientTsServerTest : TsTestBase() {
     }
 
     @Test
-    @Timeout(30, unit = TimeUnit.SECONDS)
+    @Timeout(20, unit = TimeUnit.SECONDS)
     fun multipleClients() = runBlocking(Dispatchers.IO) {
         useClient { client1 ->
             useClient { client2 ->
