+++ integration-test/src/jvmTest/kotlin/io/modelcontextprotocol/kotlin/sdk/integration/typescript/stdio/OldSchemaKotlinClientTsServerTestStdio.kt
@@ -3,9 +3,7 @@ package io.modelcontextprotocol.kotlin.sdk.integration.typescript.stdio
 import io.modelcontextprotocol.kotlin.sdk.client.Client
 import io.modelcontextprotocol.kotlin.sdk.integration.typescript.OldSchemaAbstractKotlinClientTsServerTest
 import io.modelcontextprotocol.kotlin.sdk.integration.typescript.TransportKind
-import io.modelcontextprotocol.kotlin.test.utils.DisabledOnCI
 
-@DisabledOnCI
 class OldSchemaKotlinClientTsServerTestStdio : OldSchemaAbstractKotlinClientTsServerTest() {
 
     override val transportKind = TransportKind.STDIO
