+++ integration-test/src/jvmTest/kotlin/io/modelcontextprotocol/kotlin/sdk/integration/typescript/stdio/myClient.ts
@@ -16,7 +16,7 @@ async function main() {
 
     const path = await import('path');
     const { pathToFileURL } = await import('url');
-    const clientUrl = pathToFileURL(path.join(sdkDir, 'src', 'client', 'index.ts')).href;
+    const clientUrl = pathToFileURL(path.join(sdkDir, 'packages', 'client', 'src', 'index.ts')).href;
     ({ Client } = await import(clientUrl));
 
     if (!toolName) {
