+++ integration-test/src/jvmTest/typescript/package.json
@@ -1,0 +1,26 @@
+{
+  "name": "mcp-typescript-test-sdk",
+  "version": "1.0.0",
+  "private": true,
+  "scripts": {
+    "server:stdio": "tsx server/stdio-server.ts",
+    "server:sse": "tsx server/sse-server.ts"
+  },
+  "dependencies": {
+    "@modelcontextprotocol/sdk": "^1.25.2",
+    "body-parser": "^2.2.2",
+    "express": "^5.2.1",
+    "uuid": "^13.0.0",
+    "zod": "^4.3.5"
+  },
+  "devDependencies": {
+    "@eslint/js": "^9.39.2",
+    "@types/cors": "^2.8.19",
+    "@types/express": "^5.0.6",
+    "@types/node": "^24.10.6",
+    "@types/uuid": "^10.0.0",
+    "cors": "^2.8.5",
+    "tsx": "^4.21.0",
+    "typescript": "^5.9.3"
+  }
+}
